Added schema definitions for csv files using namedtuples.
Rates are represented as Decimal() instead of float() to support strict fixed point equality invariance during sorting.

A context manager is used to read csv files to make sure that the file will be closed if any error occurs while reading the file.

Added comments to RateFinder.dump_cache_csv.

Made the <cache.csv> an optional arg so that caching is not used by default.

Added instructions to the COMMENTS to explain how to use the caching
feature.

Updated the test suite to work with Decimal values.